src/requirements.txt
# Core dependencies
gtts==2.5.1
deep-translator==1.11.4

# API Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0

# Utilities
python-multipart==0.0.9
aiofiles==24.1.0

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pydub==0.25.1

# Database & Auth (for Phase 2)
supabase==2.7.4
gotrue==2.8.1
httpx==0.27.0
websockets==13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Rate limiting
slowapi==0.1.9